seperate jframe
