+++ src/thaichessui/MenuPanel.java
@@ -1,0 +1,143 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package thaichessui;
+
+/**
+ *
+ * @author butus
+ */
+public class MenuPanel extends javax.swing.JPanel {
+
+    /**
+     * Creates new form MenuPanel
+     */
+    public MenuPanel() {
+        initComponents();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jLabel1 = new javax.swing.JLabel();
+        HostButton = new javax.swing.JButton();
+        NameField = new javax.swing.JTextField();
+        JoinButton = new javax.swing.JButton();
+        QuitButton = new javax.swing.JButton();
+
+        setBackground(new java.awt.Color(245, 255, 252));
+
+        jLabel1.setFont(new java.awt.Font("Silom", 0, 48)); // NOI18N
+        jLabel1.setForeground(new java.awt.Color(249, 186, 62));
+        jLabel1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel1.setText("MakRuk Thai");
+
+        HostButton.setFont(new java.awt.Font("Silom", 0, 18)); // NOI18N
+        HostButton.setForeground(new java.awt.Color(255, 153, 51));
+        HostButton.setText("Host");
+        HostButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                HostButtonActionPerformed(evt);
+            }
+        });
+
+        NameField.setText("Name");
+        NameField.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                NameFieldActionPerformed(evt);
+            }
+        });
+
+        JoinButton.setFont(new java.awt.Font("Silom", 0, 18)); // NOI18N
+        JoinButton.setForeground(new java.awt.Color(255, 153, 51));
+        JoinButton.setText("Join");
+        JoinButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                JoinButtonActionPerformed(evt);
+            }
+        });
+
+        QuitButton.setFont(new java.awt.Font("Silom", 0, 18)); // NOI18N
+        QuitButton.setForeground(new java.awt.Color(255, 0, 0));
+        QuitButton.setText("Quit");
+        QuitButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                QuitButtonActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
+        this.setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addContainerGap(202, Short.MAX_VALUE)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addComponent(HostButton, javax.swing.GroupLayout.PREFERRED_SIZE, 212, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(NameField, javax.swing.GroupLayout.PREFERRED_SIZE, 212, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(JoinButton, javax.swing.GroupLayout.PREFERRED_SIZE, 212, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(QuitButton, javax.swing.GroupLayout.PREFERRED_SIZE, 212, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addGap(252, 252, 252))
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
+                        .addComponent(jLabel1)
+                        .addGap(194, 194, 194))))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addGap(64, 64, 64)
+                .addComponent(jLabel1)
+                .addGap(49, 49, 49)
+                .addComponent(NameField, javax.swing.GroupLayout.PREFERRED_SIZE, 34, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(22, 22, 22)
+                .addComponent(HostButton, javax.swing.GroupLayout.PREFERRED_SIZE, 59, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addComponent(JoinButton, javax.swing.GroupLayout.PREFERRED_SIZE, 59, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(18, 18, 18)
+                .addComponent(QuitButton, javax.swing.GroupLayout.PREFERRED_SIZE, 60, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(41, Short.MAX_VALUE))
+        );
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void HostButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_HostButtonActionPerformed
+        HostPanel hostPanel = new HostPanel();
+        this.setLayout(new java.awt.BorderLayout());
+        this.removeAll();
+        this.add(hostPanel);
+        this.revalidate();
+        //TODO Open host page
+    }//GEN-LAST:event_HostButtonActionPerformed
+
+    private void NameFieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_NameFieldActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_NameFieldActionPerformed
+
+    private void JoinButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_JoinButtonActionPerformed
+        this.setVisible(false);
+        //TODO Open Join Page
+    }//GEN-LAST:event_JoinButtonActionPerformed
+
+    private void QuitButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_QuitButtonActionPerformed
+        System.exit(0);
+        
+    }//GEN-LAST:event_QuitButtonActionPerformed
+
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton HostButton;
+    private javax.swing.JButton JoinButton;
+    private javax.swing.JTextField NameField;
+    private javax.swing.JButton QuitButton;
+    private javax.swing.JLabel jLabel1;
+    // End of variables declaration//GEN-END:variables
+}
